requirements.txt
requests==2.32.3
requests_toolbelt==1.0.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
gradio
python-dotenv